updated pkgs and separated utils functions
