+++ package.json
@@ -11,7 +11,7 @@
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.20.1",
-    "chatgpt": "^4.0.3",
+    "chatgpt": "^4.1.1",
     "cors": "^2.8.5",
     "dotenv-safe": "^8.2.0",
     "ejs": "^3.1.8",
